Fixed a Visual Studio Compiler problem with sqrt

+++ src/nodes.cpp
@@ -41,12 +41,12 @@ QPoint BaseNode::getClosestConnectionPoint(const QPoint &point)
 
     double minlength, testlength;
     int minindex;
-    minlength = sqrt((point.x() - connectionPoints.at(0).x()) * (point.x() - connectionPoints.at(0).x()) +
+    minlength = sqrt((double)(point.x() - connectionPoints.at(0).x()) * (point.x() - connectionPoints.at(0).x()) +
                      (point.y() - connectionPoints.at(0).y()) * (point.y() - connectionPoints.at(0).y()));
     minindex = 0;
 
     for (int i=1; i<(int)connectionPoints.size(); i++) {
-        testlength = sqrt((point.x() - connectionPoints.at(i).x()) * (point.x() - connectionPoints.at(i).x()) +
+        testlength = sqrt((double)(point.x() - connectionPoints.at(i).x()) * (point.x() - connectionPoints.at(i).x()) +
                           (point.y() - connectionPoints.at(i).y()) * (point.y() - connectionPoints.at(i).y()));
         if (testlength < minlength) {
             minlength = testlength;
@@ -57,9 +57,8 @@ QPoint BaseNode::getClosestConnectionPoint(const QPoint &point)
     return connectionPoints.at(minindex);
 }
 
-list<BaseNode *> BaseNode::getConnectedNodes()
+list<BaseNode*> BaseNode::getConnectedNodes()
 {
-    //objects = connectedObjects;
     return list<BaseNode*>(connectedObjects);
 }
